requirements-enhanced.txt
# Enhanced scraper dependencies
# Install with: pip install -r requirements-enhanced.txt

# Document processing
pypdf>=4.0.0  # PDF text extraction
python-docx>=1.0.0  # Word document extraction
openpyxl>=3.1.0  # Excel extraction

# OCR support (optional)
pytesseract>=0.3.10  # Tesseract OCR wrapper
pdf2image>=1.16.0  # PDF to image conversion for OCR
Pillow>=10.0.0  # Image processing

# Translation (optional)
openai>=1.0.0  # OpenAI API for translation

# All existing dependencies from pyproject.toml should still be installed
# This file only adds new optional dependencies for enhanced features
